🔨refactor: Use scala.jdk.CollectionConverters instead of scala.collection.JavaConverters
