+++ src/main/scala/lerna/akka/entityreplication/raft/RaftSettings.scala
@@ -5,7 +5,7 @@ import java.util.concurrent.TimeUnit.NANOSECONDS
 import com.typesafe.config.Config
 import lerna.akka.entityreplication.util.JavaDurationConverters._
 
-import scala.collection.JavaConverters._
+import scala.jdk.CollectionConverters._
 import scala.concurrent.duration.{ Duration, FiniteDuration }
 import scala.util.Random
